pyproject.toml
[tool.poetry]
name = "aviation-topic-pipeline"
version = "0.1.0"
description = "Thematic modeling of posts from aviation Telegram channels"
authors = ["Aleksandr Tokmashov <atokmasov.71@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.15"
fastapi = ">=0.115.8,<0.116.0"
uvicorn = {extras = ["standard"], version = ">=0.35.0,<0.36.0"}
pydantic-settings = ">=2.12.0,<3.0.0"
pydantic = ">=2.10.4,<3.0.0"
python-multipart = ">=0.0.20,<0.0.21"
numpy = ">=2.2.3,<2.3.0"
scikit-learn = ">=1.7.2,<1.8.0"
pandas = ">=2.2.3,<2.3.0"
nltk = ">=3.9.2,<4.0.0"
pymorphy3 = ">=2.0.6,<3.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
jupyter = ">=1.0.0,<2.0.0"
ipykernel = ">=7.1.0,<8.0.0"
seaborn = ">=0.13.2,<0.14.0"
matplotlib = ">=3.9.2,<3.10.0"
nbformat = ">=5.10.4,<6.0.0"
pytest = ">=8.3.4,<9.0.0"
pytest-asyncio = ">=0.24.0,<0.25.0"
httpx = ">=0.28.1,<0.29.0"
mlflow = ">=3.6.0,<4.0.0"
dvc = ">=3.64.0,<4.0.0"
dvc-gdrive = ">=3.0.1,<4.0.0"
wordcloud = ">=1.9.4,<2.0.0"
telethon = ">=1.42.0,<2.0.0"

[tool.poetry.group.ml.dependencies]
hdbscan = ">=0.8.40,<0.9.0"
umap-learn = ">=0.5.9.post2,<0.6.0"
transformers = ">=4.57.3,<5.0.0"
triton = ">=3.5.1,<4.0.0"
bertopic = ">=0.17.3,<0.18.0"


[tool.poetry.scripts]
aviation-api = "app.main:app"
